requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
python-dotenv>=1.0.0
rich>=13.0.0
streamlit>=1.35.0
